fixed return type of the NetworkInterface method getMTU() in all its OS implementations

+++ oshi-core/src/main/java/oshi/hardware/NetworkIF.java
@@ -111,7 +111,7 @@ public interface NetworkIF {
      *         This value is set when the {@link oshi.hardware.NetworkIF} is
      *         instantiated and may not be up to date.
      */
-    int getMTU();
+    long getMTU();
 
     /**
      * The Media Access Control (MAC) address.
@@ -402,8 +402,7 @@ public interface NetworkIF {
         /**
          * Find IfOperStatus by the integer value.
          *
-         * @param value
-         *            Integer value specified in RFC 2863
+         * @param value Integer value specified in RFC 2863
          * @return the matching IfOperStatu or UNKNOWN if no matching IfOperStatus can
          *         be found
          */
